Separate out some Play page utility functions
