+++ src/views/Play/utils.ts
@@ -24,3 +24,57 @@ export function anyChallengesToShow(filter: ChallengeFilter, challenge_list: Cha
         }, false)
     );
 }
+export function challenge_sort(A: Challenge, B: Challenge) {
+    if (A.eligible && !B.eligible) {
+        return -1;
+    }
+    if (!A.eligible && B.eligible) {
+        return 1;
+    }
+
+    if (A.user_challenge && !B.user_challenge) {
+        return -1;
+    }
+    if (!A.user_challenge && B.user_challenge) {
+        return 1;
+    }
+
+    const t = A.username.localeCompare(B.username);
+    if (t) {
+        return t;
+    }
+
+    if (A.ranked && !B.ranked) {
+        return -1;
+    }
+    if (!A.ranked && B.ranked) {
+        return 1;
+    }
+
+    return A.challenge_id - B.challenge_id;
+}
+
+export function time_per_move_challenge_sort(A: Challenge, B: Challenge) {
+    const comparison = Math.sign(A.time_per_move - B.time_per_move);
+
+    if (comparison) {
+        return comparison;
+    }
+
+    if (A.eligible && !B.eligible) {
+        return -1;
+    }
+    if (!A.eligible && B.eligible) {
+        return 1;
+    }
+    if (A.user_challenge && !B.user_challenge) {
+        return -1;
+    }
+    if (!A.user_challenge && B.user_challenge) {
+        return 1;
+    }
+
+    const createdA = A.created ? new Date(A.created).getTime() : -Infinity;
+    const createdB = B.created ? new Date(B.created).getTime() : -Infinity;
+    return createdA - createdB;
+}
